made model list CL-configurable in run_hsw script

+++ scripts/run_hsw.sh
@@ -9,20 +9,25 @@
 
 set -e
 
+# Options
+clean=false
+models="alexnet vgg11 resnet50 inceptionV3 lstm cnn3d"
+if [ $# -ge 1 ]; then models=$@; fi
+
 # Configuration
 export OMP_NUM_THREADS=32
 export KMP_AFFINITY="granularity=fine,compact,1,0"
 export KMP_BLOCKTIME=1
 export BENCHMARK_RESULTS_PATH=$SCRATCH/pytorch-benchmarks/hsw-v1.2.0-n${SLURM_JOB_NUM_NODES}
-[ -d $BENCHMARK_RESULTS_PATH ] && rm -rf $BENCHMARK_RESULTS_PATH
+if $clean; then
+    [ -d $BENCHMARK_RESULTS_PATH ] && rm -rf $BENCHMARK_RESULTS_PATH
+fi
 module load pytorch/v1.2.0
 
-srun -l python train.py -d configs/alexnet.yaml
-srun -l python train.py -d configs/vgg11.yaml
-srun -l python train.py -d configs/resnet50.yaml
-srun -l python train.py -d configs/inceptionV3.yaml
-srun -l python train.py -d configs/lstm.yaml
-srun -l python train.py -d configs/cnn3d.yaml
+# Run each model
+for m in $models; do
+    srun -l python train.py -d configs/${m}.yaml
+done
 
 echo "Collecting benchmark results..."
 python parse.py $BENCHMARK_RESULTS_PATH -o $BENCHMARK_RESULTS_PATH/results.txt
